+++ example/lib/player/knight.dart
@@ -274,7 +274,18 @@ class Knight extends SimplePlayer with Lighting, ObjectCollision, MouseGesture {
           gameRef.context,
           [
             Say(
-              "Look at this! It seems that I'm not alone here ...",
+              text: [
+                TextSpan(
+                  text: "Look at this! It seems that",
+                ),
+                TextSpan(
+                  text: ' I\'m not alone ',
+                  style: TextStyle(color: Colors.red),
+                ),
+                TextSpan(
+                  text: "here...",
+                ),
+              ],
               person: Container(
                 width: 100,
                 height: 100,
@@ -292,7 +303,7 @@ class Knight extends SimplePlayer with Lighting, ObjectCollision, MouseGesture {
             ),
           ],
           finish: () {
-            print('finish');
+            print('finish talk');
             gameRef.camera.moveToPlayerAnimated();
           },
           logicalKeyboardKeyToNext: LogicalKeyboardKey.space,